Add files via upload
Image  inpainting is a art of missing value or a data in an image. The purpose of image inpainting is to reconstruct missing regions
which is visible for human eyes. Image inpainting is the process of reconstructing lost part of images based on the background
information. Image inpainting is a technique for restoring damaged old photographs and removing undesired objects from an image.
The basic idea behind the technique is to automatically fill in lost or missing parts of an image using information from the
surrounding area. It is used for restoration of old films and object removal in digital photographs. It is also applied to red-eye
correction, stamped data from photographs, dust spot in film, removing objects to creative effect etc. The main goal of the Inpainting
algorithm is to modify the damaged region in an image. In this paper we provide a review of different techniques used for image
Inpainting. We discuss texture synthesis method and inpaint the image using masking.